on free time delete, remove booked times that are no longer valid

+++ app/routes/free_time.py
@@ -1,5 +1,6 @@
 from flask import Blueprint, Response, json, request
 from app.services.mongo_service import MongoService
+from app.services.booking_service import BookingService
 
 
 ft_db = MongoService(
@@ -10,6 +11,16 @@ ft_db = MongoService(
 )
 
 free_time_bp = Blueprint("free-time", __name__)
+
+booking_db = MongoService(
+    {
+        "database": "scheduler",
+        "collection": "booking",
+    }
+)
+
+
+booking_service = BookingService(ft_db, booking_db)
 
 
 @free_time_bp.route("/free-time", methods=["GET"])
@@ -94,7 +105,19 @@ def update_free_time(consultant_id, id):
 
 @free_time_bp.route("/consultant/<consultant_id>/free-time/<id>", methods=["DELETE"])
 def delete_free_time(consultant_id, id):
+    free_time_to_delete = ft_db.read(id, {"consultant_id": consultant_id})
+    if free_time_to_delete is None:
+        return Response(
+            status=404,
+            mimetype="application/json",
+        )
+
     response = ft_db.delete({"id": id, "consultant_id": consultant_id})
+
+    # # TODO: delete all bookings
+    print({"free_time_to_delete": free_time_to_delete})
+    booking_service.cancel_bookings(cancelled_free_time=free_time_to_delete)
+
     return Response(
         response=json.dumps(response), status=200, mimetype="application/json"
     )
